+++ lib/hpfeeds.rb
@@ -2,6 +2,8 @@ require 'hpfeeds/version'
 
 module HPFeeds
   autoload :Client, 'hpfeeds/client'
-  autoload :Exception, 'hpfeeds/exception'
   autoload :Decoder, 'hpfeeds/decoder'
+  # exceptions
+  autoload :Exception, 'hpfeeds/exception'
+  autoload :ErrorMessage, 'hpfeeds/exception'
 end