+++ lib/hpfeeds/exception.rb
@@ -1,3 +1,4 @@
 module HPFeeds
   class Exception < RuntimeError; end
+  class ErrorMessage < RuntimeError; end
 end