+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### current
 - major fixes:
  - separate handlers for messages from different chanels
+ - better error messages handling
 
 - minor fixes:
  - documentation, indentation